# openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorBodyOut": {
        "description": "The body of an error: a top-level code, its params, and optional per-field detail.",
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "fields": {
            "anyOf": [
              {
                "additionalProperties": {
                  "$ref": "#/components/schemas/FieldErrorOut"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fields"
          },
          "params": {
            "additionalProperties": true,
            "default": {},
            "title": "Params",
            "type": "object"
          }
        },
        "required": [
          "code"
        ],
        "title": "ErrorBodyOut",
        "type": "object"
      },
      "ErrorEnvelope": {
        "description": "`{ \"error\": { \"code\": ..., \"params\": {...}, \"fields\": {...}? } }`.",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBodyOut"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
      "FieldErrorOut": {
        "description": "One field's `{code, params}` failure.",
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "params": {
            "additionalProperties": true,
            "default": {},
            "title": "Params",
            "type": "object"
          }
        },
        "required": [
          "code"
        ],
        "title": "FieldErrorOut",
        "type": "object"
      },
...
      "LabelCreate": {
        "additionalProperties": false,
        "description": "Request body for creating a label.",
        "properties": {
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "LabelCreate",
        "type": "object"
      },
      "LabelEnvelope": {
        "description": "`{ \"label\": Label }`.",
        "properties": {
          "label": {
            "$ref": "#/components/schemas/LabelOut"
          }
        },
        "required": [
          "label"
        ],
        "title": "LabelEnvelope",
        "type": "object"
      },
      "LabelListEnvelope": {
        "description": "`{ \"labels\": [Label] }`.",
        "properties": {
          "labels": {
            "items": {
              "$ref": "#/components/schemas/LabelOut"
            },
            "title": "Labels",
            "type": "array"
          }
        },
        "required": [
          "labels"
        ],
        "title": "LabelListEnvelope",
        "type": "object"
      },
      "LabelOut": {
        "description": "Wire representation of a label.",
        "properties": {
          "color": {
            "title": "Color",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "use_count": {
            "title": "Use Count",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "name",
          "color",
          "use_count"
        ],
        "title": "LabelOut",
        "type": "object"
      },
      "LabelUpdate": {
        "additionalProperties": false,
        "description": "Request body for partially updating a label.",
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          }
        },
        "title": "LabelUpdate",
        "type": "object"
      },
      "PersonCreate": {
        "additionalProperties": false,
        "description": "Request body for creating a person on a trip.",
        "properties": {
          "color": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Color"
          },
          "default_weight": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Default Weight"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "PersonCreate",
        "type": "object"
      },
      "PersonEnvelope": {
        "description": "`{ \"person\": Person }`.",
        "properties": {
          "person": {
            "$ref": "#/components/schemas/PersonOut"
          }
        },
        "required": [
          "person"
        ],
        "title": "PersonEnvelope",
        "type": "object"
      },
      "PersonListEnvelope": {
        "description": "`{ \"people\": [Person] }`.",
        "properties": {
          "people": {
            "items": {
              "$ref": "#/components/schemas/PersonOut"
            },
            "title": "People",
            "type": "array"
          }
        },
        "required": [
          "people"
        ],
        "title": "PersonListEnvelope",
        "type": "object"
      },
      "PersonOut": {
        "description": "Wire representation of a person on a trip.",
        "properties": {
          "active": {
            "title": "Active",
            "type": "boolean"
          },
          "color": {
            "title": "Color",
            "type": "string"
          },
          "default_weight": {
            "title": "Default Weight",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "initial": {
            "title": "Initial",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "sort_order": {
            "title": "Sort Order",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "name",
          "initial",
          "color",
          "default_weight",
          "sort_order",
          "active"
        ],
        "title": "PersonOut",
        "type": "object"
      },
      "PersonUpdate": {
        "additionalProperties": false,
        "description": "Request body for partially updating a person.",
        "properties": {
          "active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active"
          },
          "default_weight": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Default Weight"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "sort_order": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sort Order"
          }
        },
        "title": "PersonUpdate",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Expense tracking for a shared trip: people, currencies, countries and labels belong to a trip; every line item records who paid and how the cost splits across the roster.\n\nJSON in, JSON out. Every response is a single object with its payload under a named key, never a bare array. Amounts are plain numbers in the currency's own units - typed values to 2 fraction digits, computed shares to 6 - and are taken on input as canonical decimal strings (`\"18400.50\"`). Nothing is ever converted between currencies: balances and statistics are per currency.\n\nErrors are `{code, params}` with no message text, so a client renders them through its own catalog in its own language.",
    "title": "Trip Accounter API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/trips/{slug}/labels": {
      "get": {
        "description": "List a trip's labels, most used first.",
        "operationId": "list_labels_api_v1_trips__slug__labels_get",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LabelListEnvelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "List Labels",
        "tags": [
          "labels"
        ]
      },
      "post": {
        "description": "Create a new label on a trip.",
        "operationId": "create_label_api_v1_trips__slug__labels_post",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LabelCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LabelEnvelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Create Label",
        "tags": [
          "labels"
        ]
      }
    },
    "/api/v1/trips/{slug}/labels/{label_id}": {
      "delete": {
        "description": "Delete a label from a trip.",
        "operationId": "delete_label_api_v1_trips__slug__labels__label_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "label_id",
            "required": true,
            "schema": {
              "title": "Label Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Delete Label",
        "tags": [
          "labels"
        ]
      },
      "patch": {
        "description": "Rename a label.",
        "operationId": "update_label_api_v1_trips__slug__labels__label_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "label_id",
            "required": true,
            "schema": {
              "title": "Label Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LabelUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LabelEnvelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Update Label",
        "tags": [
          "labels"
        ]
      }
    },
    "/api/v1/trips/{slug}/people": {
      "get": {
        "description": "List people on a trip, in sort order.",
        "operationId": "list_people_api_v1_trips__slug__people_get",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonListEnvelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "List People",
        "tags": [
          "people"
        ]
      },
      "post": {
        "description": "Add a new person to a trip's roster.",
        "operationId": "create_person_api_v1_trips__slug__people_post",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PersonCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonEnvelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Create Person",
        "tags": [
          "people"
        ]
      }
    },
    "/api/v1/trips/{slug}/people/{person_id}": {
      "delete": {
        "description": "Delete a person from the roster.",
        "operationId": "delete_person_api_v1_trips__slug__people__person_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "person_id",
            "required": true,
            "schema": {
              "title": "Person Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Delete Person",
        "tags": [
          "people"
        ]
      },
      "patch": {
        "description": "Update a person's fields.",
        "operationId": "update_person_api_v1_trips__slug__people__person_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "person_id",
            "required": true,
            "schema": {
              "title": "Person Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PersonUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonEnvelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Update Person",
        "tags": [
          "people"
        ]
      }
    },
...
  }
}